Add `verifyTrace` buildImage option
This option can to be set to debug digest mismatch issue. It generates
a trace a build time and the script `image.verifyTrace`. These traces
contains all file attribute written to the tar stream and the chechsum
of all file writter to the tar stream.

This could allow to identify differences because of the sandbox used
at build time while the Nix build sandbox is not used to run time,
when the image is pushed to a destination.